BE-012(Feat): crear and get Appointments

+++ Appointments.Application/DTO/AppointmentsDto.cs
@@ -3,6 +3,6 @@
 namespace Appointment.Application.DTO
 {
     public record AppointmentsDto(int appointmentId, int userId, int petId, DateOnly appointmentDate
-        , StatusAppointments statusAppointments, float totalPrice, PaymentStatus paymentStatus
+        , StatusAppointments statusAppointments, double totalPrice, PaymentStatus paymentStatus
         , PaymentMethod paymentMethod, string notes);
 }
