+++ PetShop.Domain/Entities/Appointments.cs
@@ -1,7 +1,6 @@
 ﻿using PetShop.Domain.Entities.Enums;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.ComponentModel.DataAnnotations;
-using PetShop.Domain.Entities;
 
 
 namespace PetShop.Domain.Entities
@@ -21,7 +20,7 @@ namespace PetShop.Domain.Entities
 
         public DateTime AppointmentDate { get; set; }
         public StatusAppointments StatusAppointment { get; set; }
-        public float TotalPrice { get; set; }
+        public double TotalPrice { get; set; }
         public PaymentStatus PaymentStatus { get; set; }
         public PaymentMethod PaymentMethod { get; set; }
         public string Notes { get; set; }
